backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
scikit-learn
sentence-transformers
pydantic
pydantic[email]
aiofiles
numpy
pandas
python-dotenv
firebase-admin
google-generativeai
PyPDF2
python-docx
docx2txt
pydantic-settings